fix(reporter): multiple minors and cleanup
relates-to #45
