+++ packages/reporter/src/index.ts
@@ -1,1 +1,2 @@
-export * from './lib/Reporter';
+export { Reporter } from './lib';
+export { StdReporter } from './std';
