+++ packages/reporter/src/models/Scan.ts
@@ -1,5 +1,6 @@
 import { Issue } from './Issue';
 
+// TODO utilize models from @secbox/scan
 export interface Scan {
   issues(): Promise<Issue[]>;
 }
