+++ packages/reporter/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { IssuesGrouper } from './issues-grouper';
